# PROTON-1965 Optimize CompositeReadableBuffer::equals with single chunk