Using the single chunk directly while performing the byte comparison increase the performance.
Master:
```
Benchmark                                                                      (chunks)  (direct)  (size)  Mode  Cnt     Score    Error   Units
CompositeReadableBufferBenchmark.equalsToByteBufferReader                             1     false       8  avgt    5    18.127 ±  1.568   ns/op
CompositeReadableBufferBenchmark.equalsToByteBufferReader                             1     false      64  avgt    5    40.725 ±  2.329   ns/op
CompositeReadableBufferBenchmark.equalsToByteBufferReader                             1     false    1024  avgt    5   460.801 ± 13.786   ns/op
```
This PR:
```
Benchmark                                                                      (chunks)  (direct)  (size)  Mode  Cnt     Score     Error   Units
CompositeReadableBufferBenchmark.equalsToByteBufferReader                             1     false       8  avgt    5    27.799 ±   0.462   ns/op
CompositeReadableBufferBenchmark.equalsToByteBufferReader                             1     false      64  avgt    5   108.563 ±   1.236   ns/op
CompositeReadableBufferBenchmark.equalsToByteBufferReader                             1     false    1024  avgt    5  5251.102 ± 261.561   ns/op
```