@tabish121 2018-11-18
@franz1981 Looks good so far, would like to see the formatting changed to match the remainder of the code. The reset is a good catch, that is a bug in the current code.