@gemmellr 2018-11-19
Can you unwind the 'revert' of my currentPos -> origPos variable name change from earlier? :)