@gemmellr 2018-11-20
I've pushed an updated version of the change with a fix for the offset issue I noticed yesterday, which Tim added a test to cover.