@gemmellr 2018-11-21
For later reference, a related further change was made via b3a66b741c34b40b059fc023b3c7650720dfc001 to apply the optimisation in more cases.